skal/requirements.txt
alembic==0.9.9
astroid==1.6.4
Babel==2.5.3
blinker==1.4
click==6.7
dominate==2.3.1
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Limiter==1.0.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
limits==1.3
lipsum==0.1.2
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
PyJWT==1.6.1
pylint==1.9.0
python-dateutil==2.7.3
python-dotenv==0.8.2
python-editor==1.0.3
pytz==2018.4
six==1.11.0
SQLAlchemy==1.2.7
urllib3==1.22
visitor==0.1.3
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.1
